create knowledge giver class, create performance rateable module, delete salary_setter and add to knowledge giver class
