+++ class_helpers.rb
@@ -1,3 +1,4 @@
 require_relative 'person'
 require_relative 'phase_settable'
-require_relative 'salary_setter'
+require_relative 'performance_setter'
+require_relative 'knowledge_giver'
