+++ performance_setter.rb
@@ -1,0 +1,25 @@
+module PerformanceRateable
+  def set_performance_rating_teacher(rating)
+    response = ""
+    if rating > 90
+      receive_raise(@target_raise)
+      response = "Yay, I'm a great employee!"
+    else
+      response += "Oh, well -- thanks to this actionable, specific, and kind "
+      response += "feedback, I'll do better next time."
+    end
+    response
+  end
+
+  def set_performance_rating_apprentice(rating)
+    response = ""
+    if rating > 80
+      response = "Yay, I'm a great employee!"
+      receive_raise(@target_raise)
+    else
+      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
+    end
+    response
+  end
+
+end
