--- salary_setter.rb
@@ -1,12 +1,0 @@
-module Salaryable
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-end
