+++ teacher.rb
@@ -1,12 +1,11 @@
 require_relative 'class_helpers'
 
-class Teacher < Person
-  include PhaseSettable, Salaryable
-  attr_reader  :salary, :performance_rating, :target_raise
+class Teacher < KnowledgeGiver
+  include PhaseSettable, PerformanceRateable
+  attr_reader :performance_rating
 
   def initialize(options={})
     super
-    @phase = 3
     @target_raise = 1000
   end
 
@@ -18,15 +17,4 @@ class Teacher < Person
     response
   end
 
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      receive_raise(@target_raise)
-      response = "Yay, I'm a great employee!"
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
-  end
 end
